Fixed Author names being undefined.
